♻️ refactor [#58] : 단순 로직 수정
